+++ packages/riverpod_lint_flutter_test/test/lints/notifier_extends/notifier_extends.g.dart
@@ -10,7 +10,7 @@ part of 'notifier_extends.dart';
 // ignore_for_file: type=lint, type=warning
 
 @ProviderFor(MyNotifier)
-const myNotifierProvider = MyNotifierProvider._();
+const myProvider = MyNotifierProvider._();
 
 final class MyNotifierProvider extends $NotifierProvider<MyNotifier, int> {
   const MyNotifierProvider._()
@@ -18,7 +18,7 @@ final class MyNotifierProvider extends $NotifierProvider<MyNotifier, int> {
           from: null,
           argument: null,
           retry: null,
-          name: r'myNotifierProvider',
+          name: r'myProvider',
           isAutoDispose: true,
           dependencies: null,
           $allTransitiveDependencies: null,